[Half-Spectre] Add support for Half-Spectre Gadgets
This commit introduces a new type of gadgets, i.e. _Half-Spectre_
gadgets, alongside the existing Dispatch Gadgets and Transmission
Gadgets.

Half-Spectre gadgets are commonly just attacker-controlled loads in
a victim, that can be used to prefetch a secret in the cache,
load it into the LFB, or even for hammering.

The detection of these gadgets is off by default, but can be turned on
in the config file.